Move the code to authenticate with GEE from common.py to downloads.py
